Moved LoadIndicator to common-ui
