+++ common-ui/components/load-indicator/load-indicator.styles.tsx
